Update ADhomelab.md
After completing the lab setup, I not only have a functional Active Directory environment but also the skills to recreate it independently. This hands-on experience will familiarize me with various network configuration options and the essential tools needed to configure Windows Server 2019 and Windows 10 clients, effectively simulating an enterprise-grade environment.